+++ Models/Product.cs
@@ -17,8 +17,8 @@ namespace DPSP_Api
 
         public int Id { get; set; }
         public string Name { get; set; }
-        public double Cost { get; set; }
-        public double? Rating { get; set; }
+        public float Cost { get; set; }
+        public float? Rating { get; set; }
         public bool? Avail { get; set; }
         public int IdCategory { get; set; }
         public int IdStoreInfo { get; set; }
